Extracted app validation into new module
